+++ replacer-frontend/src/app/admin/admin-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { AuthenticationGuard } from '../core/guard/authentication.guard';
 import { AdminComponent } from './admin.component';
 
-const routes: Routes = [{ path: 'admin', component: AdminComponent, canActivate: [AuthenticationGuard] }];
+const routes: Routes = [{ path: '', component: AdminComponent, canActivate: [AuthenticationGuard] }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],